+++ package.json
@@ -19,6 +19,7 @@
     "react": "^16.13.1",
     "react-debounce-input": "^3.2.2",
     "react-dom": "^16.13.1",
+    "react-hotkeys-hook": "^2.4.0",
     "react-konva": "^16.13.0-3",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
